requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
Pillow==12.1.0
python-dotenv==1.2.1